+++ FakeServer/Controllers/DynamicController.cs
@@ -62,17 +62,23 @@ namespace FakeServer.Controllers
         /// <param name="take">Items to take (optional name limit)</param>
         /// <returns>List of items</returns>
         /// <response code="200">Collection item array</response>
+        /// <response code="400">Invalid query parameters</response>
         /// <response code="404">Collection not found</response>
         [HttpGet("{collectionId}")]
         public IActionResult GetItems(string collectionId, int skip = 0, int take = 512)
         {
-            var datas = _ds.GetCollection(collectionId).AsQueryable();
+            var collection = _ds.GetCollection(collectionId);
 
             // Collection can actually just be empty, but in this case we handle it as it is not found
-            if (!datas.Any())
+            if (!collection.AsQueryable().Any())
                 return NotFound();
 
             var options = GetQueryOptions(Request.Query, skip, take);
+
+            if (!options.Validate())
+                return BadRequest();
+
+            var datas = options.IsTextSearch ? collection.Find(Request.Query["q"]) : collection.AsQueryable();
 
             foreach (var key in options.QueryParams)
             {
@@ -263,6 +269,7 @@ namespace FakeServer.Controllers
         {
             var skipWord = "skip";
             var takeWord = "take";
+            var isTextSearch = false;
 
             var queryParams = query.Keys.ToList();
 
@@ -280,10 +287,16 @@ namespace FakeServer.Controllers
                 takeWord = "limit";
             }
 
+            if (queryParams.Contains("q"))
+            {
+                isTextSearch = true;
+                queryParams.Remove("q");
+            }
+
             queryParams.Remove("skip");
             queryParams.Remove("take");
 
-            return new QueryOptions { Skip = skip, Take = take, SkipWord = skipWord, TakeWord = takeWord, QueryParams = queryParams };
+            return new QueryOptions { Skip = skip, Take = take, SkipWord = skipWord, TakeWord = takeWord, IsTextSearch = isTextSearch, QueryParams = queryParams };
         }
     }
 
@@ -297,6 +310,16 @@ namespace FakeServer.Controllers
 
         public string TakeWord { get; set; }
 
+        public bool IsTextSearch { get; set; }
+
         public List<string> QueryParams { get; set; }
+
+        public bool Validate()
+        {
+            if (IsTextSearch && QueryParams.Any())
+                return false;
+
+            return true;
+        }
     }
 }