+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
 ### [Unreleased]
+*
+
+### [0.3.0] - 2017-08-17
 * Offset and limit slice option
 * Option for query to return JSON object, instead of the results array
 * Full-text search
+* Experimental GraphQL query support
 
 ### [0.2.0] - 2017-07-31
 * Get Controller's route template from constant value